kalip_erp/backend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Cython==0.29.28
Django==4.0.3
django-bootstrap5==21.3
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-graphql-jwt==0.3.4
django-js-asset==2.0.0
django-modeladmin-reorder==0.3.1
django-neomodel==0.0.7
djangorestframework-queryfields==1.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
Pillow==9.0.1
promise==2.3
PyJWT==2.3.0
pytz==2021.3
Rx==1.6.1
Shapely==1.7.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
